Fix iteration order docs
The iteration order was wrongly described as LRU when in fact the
iteration happens in MRU order and LRU order is expensive to establish
in the current implementation.

To iterate in the opposite direction (and thus allow popping the least
recently used item when the cache is not full), one would have to keep
explicit track of where that item lives - feasible, but the cost of
keeping it up to date would have to be considered.